modify for loop of field generation

+++ source/give_clue.py
@@ -14,13 +14,14 @@ import sys
 class Clue(object):
     """ Class for a clue with its score and ranked (card, score) answer pairs."""
 
-    def __init__(self, clue, sorted_card_score_pairs, delta, penalize_negative, alpha, team, logger):
+    def __init__(self, clue, sorted_card_score_pairs, delta, penalize_negative, normalize_negative, alpha, team, logger):
         self.clue = clue
         self.sorted_card_score_pairs = sorted_card_score_pairs    
         self.team = team
         self.logger = logger
         self.delta = delta
         self.penalize_negative = penalize_negative
+        self.normalize_negative = normalize_negative
         self.alpha = alpha
         self.cropped_threshold = None
         self.total_score, self.clue_number = self._calculate_score_with_threshold()
@@ -38,6 +39,7 @@ class Clue(object):
 
         clue_number = 0
         positive_score, negative_score = 0, 0
+        negative_number = 0
         total_score = 0
 
         # find largest negative score
@@ -56,6 +58,7 @@ class Clue(object):
                positive_score += score
            elif card.color not in [self.team, "DOUBLE"]:
                negative_score += score
+               negative_number += 1
            else:
                continue
 
@@ -77,7 +80,10 @@ class Clue(object):
                     positive_score = 0
                 break
 
-        total_score = (1-self.alpha) * positive_score - self.alpha * negative_score
+        if self.normalize_negative:
+            total_score = (1-self.alpha) * positive_score - self.alpha * negative_score / negative_number
+        else:
+            total_score = (1-self.alpha) * positive_score - self.alpha * negative_score
         self.logger.debug("word: {}, positive_score: {}, negative_score: {}, total_score: {}".format(self.clue, positive_score, negative_score, total_score))
         return total_score, clue_number
     
@@ -113,9 +119,13 @@ class Spymaster(object):
     :param logger: logger for the spymaster.
     """
 
-    def __init__(self, field, embeddings, vocabulary_path, similarities_table_path, logger):
-        
-        self.similarities_table_path = similarities_table_path
+    def __init__(self, field, embeddings, vocabulary_path, keep_similarities_table, similarities_table_path,
+                 logger, game_count):
+
+        self.game_count = game_count
+        self.keep_similarities_table = keep_similarities_table
+        if self.keep_similarities_table:
+            self.similarities_table_path = similarities_table_path
         self.vocabulary_path = vocabulary_path
         self.field = field
         self.logger = logger     
@@ -192,8 +202,8 @@ class Spymaster(object):
         
         similarities_table_path = self.similarities_table_path
 
-        if similarities_table_path and os.path.exists(similarities_table_path):
-            raise ValueError("self.similarities_table_path is deprecated now. Do not give similarities_table value.")
+        if self.keep_similarities_table and similarities_table_path and os.path.exists(similarities_table_path):
+            # raise ValueError("self.similarities_table_path is deprecated now. Do not give similarities_table value.")
             self.logger.info(similarities_table_path + " exists. Loading.")
             with open(similarities_table_path, 'rb') as r:
                 self.similarities_table = pickle.load(r)
@@ -215,11 +225,14 @@ class Spymaster(object):
                         self.similarities_table[w_ix][c_ix] = similarity
                         # self.similarities_table[w_ix][c_ix] = self.model.similarity(word, card.name)
             
-            if (similarities_table_path):
+            if similarities_table_path and self.keep_similarities_table:
+                if os.path.exists(similarities_table_path):
+                    raise ValueError("{} already exists, but tries to save it.".format(similarities_table_path))
                 with open(similarities_table_path, 'wb') as w:
+                    self.logger.info("saveing table to {}...".format(similarities_table_path))
                     pickle.dump(self.similarities_table, w)
 
-    def give_clue_with_threshold(self, team, turn_count, delta, penalize_negative, alpha, top_to_print=5):
+    def give_clue_with_threshold(self, team, turn_count, delta, penalize_negative, normalize_negative, alpha, top_to_print=5):
         """
         Give a clue, maximizig the sum of similarities to the positive words set
         and minimizing the average snimilarity to all negative words of the field.
@@ -238,6 +251,7 @@ class Spymaster(object):
             raise ValueError("Team string must be RED or BLUE.")
        
         clue_candidates = []
+        is_hacky_clue = False
         for clue in self.vocab:
             card_score_pairs = []
             clue_ix = self.vocab[clue].index
@@ -246,23 +260,29 @@ class Spymaster(object):
                 card_score_pairs.append((card, score))
 
             sorted_card_score_pairs = sorted(card_score_pairs, key=lambda x: x[1], reverse=True)
-            clue = Clue(clue, sorted_card_score_pairs, delta, penalize_negative, alpha, team, self.logger)
+            clue = Clue(clue, sorted_card_score_pairs, delta, penalize_negative, normalize_negative, alpha, team, self.logger)
             clue_candidates.append(clue)
 
 
         clue_candidates = sorted(clue_candidates, key=lambda x: x.total_score, reverse=True)
 
-        # drop the clue from the candidates, whose clue_num is equal to 0 .
+        # exclude the clue whose clue_num is equal to 0 .
         clue_candidates = list(filter(lambda x: x.clue_number > 0, clue_candidates))
 
-        # find the clue whose threshold is equal to hparam: alpha.
-        clue = list(filter(lambda x: x.cropped_threshold is None, clue_candidates))[0]
+        # find the clue whose threshold is equal to (more than) hparam: alpha.
+        uncropped_clue_candidates = list(filter(lambda x: x.cropped_threshold is None, clue_candidates))
 
-        if clue is None:
+        # if there was no clue (it happens when threshold is too high), pick hacky clues.
+        if len(uncropped_clue_candidates) == 0:
             clue = list(filter(lambda x: x.cropped_threshold is not None, clue_candidates))[0]
+            is_hacky_clue = True
             self.logger.info("none of the clue similarity has exceeded threshold. clue: {} has chosen,".format(clue.clue))
-        
-        for clue in clue_candidates[:top_to_print]:
-            self.logger.info(clue.get_summary())
 
-        return clue
+        else:
+            clue = uncropped_clue_candidates[0]
+
+        # for clue in clue_candidates[:top_to_print]:
+        #     self.logger.info(clue.get_summary())
+        self.logger.info(clue.get_summary())
+
+        return clue, is_hacky_clue
